# openapi.yaml
openapi: 3.1.0
info:
  title: Human Design API
  version: 1.5.0
paths:
  /calculate:
    get:
      summary: Calculate Hd
      operationId: calculate_hd_calculate_get
      security:
      - HTTPBearer: []
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
          description: Birth year
          title: Year
        description: Birth year
      - name: month
        in: query
        required: true
        schema:
          type: integer
          description: Birth month
          title: Month
        description: Birth month
      - name: day
        in: query
        required: true
        schema:
          type: integer
          description: Birth day
          title: Day
        description: Birth day
      - name: hour
        in: query
        required: true
        schema:
          type: integer
          description: Birth hour
          title: Hour
        description: Birth hour
      - name: minute
        in: query
        required: true
        schema:
          type: integer
          description: Birth minute
          title: Minute
        description: Birth minute
      - name: second
        in: query
        required: false
        schema:
          type: integer
          description: Birth second (optional, default 0)
          default: 0
          title: Second
        description: Birth second (optional, default 0)
      - name: place
        in: query
        required: true
        schema:
          type: string
          description: Birth place (city, country)
          title: Place
        description: Birth place (city, country)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /bodygraph:
    get:
      summary: Get Bodygraph Image
      operationId: get_bodygraph_image_bodygraph_get
      security:
      - HTTPBearer: []
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
          description: Birth year
          title: Year
        description: Birth year
      - name: month
        in: query
        required: true
        schema:
          type: integer
          description: Birth month
          title: Month
        description: Birth month
      - name: day
        in: query
        required: true
        schema:
          type: integer
          description: Birth day
          title: Day
        description: Birth day
      - name: hour
        in: query
        required: true
        schema:
          type: integer
          description: Birth hour
          title: Hour
        description: Birth hour
      - name: minute
        in: query
        required: true
        schema:
          type: integer
          description: Birth minute
          title: Minute
        description: Birth minute
      - name: second
        in: query
        required: false
        schema:
          type: integer
          description: Birth second (optional, default 0)
          default: 0
          title: Second
        description: Birth second (optional, default 0)
      - name: place
        in: query
        required: true
        schema:
          type: string
          description: Birth place (city, country)
          title: Place
        description: Birth place (city, country)
      - name: fmt
        in: query
        required: false
        schema:
          type: string
          description: 'Image format: png, svg, jpg/jpeg'
          default: png
          title: Fmt
        description: 'Image format: png, svg, jpg/jpeg'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /transits/solar_return:
    get:
      summary: Get Solar Return
      operationId: get_solar_return_transits_solar_return_get
      security:
      - HTTPBearer: []
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
          description: Birth year
          title: Year
        description: Birth year
      - name: month
        in: query
        required: true
        schema:
          type: integer
          description: Birth month
          title: Month
        description: Birth month
      - name: day
        in: query
        required: true
        schema:
          type: integer
          description: Birth day
          title: Day
        description: Birth day
      - name: hour
        in: query
        required: true
        schema:
          type: integer
          description: Birth hour
          title: Hour
        description: Birth hour
      - name: minute
        in: query
        required: false
        schema:
          type: integer
          description: Birth minute (default 0)
          default: 0
          title: Minute
        description: Birth minute (default 0)
      - name: second
        in: query
        required: false
        schema:
          type: integer
          description: Birth second (optional, default 0)
          default: 0
          title: Second
        description: Birth second (optional, default 0)
      - name: place
        in: query
        required: true
        schema:
          type: string
          description: Birth place (city, country)
          title: Place
        description: Birth place (city, country)
      - name: sr_year_offset
        in: query
        required: false
        schema:
          type: integer
          description: Calculate Solar Return for X years after birth. 0 is the current
            SR.
          default: 0
          title: Sr Year Offset
        description: Calculate Solar Return for X years after birth. 0 is the current
          SR.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /transits/daily:
    get:
      summary: Get Daily Transit
      operationId: get_daily_transit_transits_daily_get
      security:
      - HTTPBearer: []
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
          description: Birth year
          title: Year
        description: Birth year
      - name: month
        in: query
        required: true
        schema:
          type: integer
          description: Birth month
          title: Month
        description: Birth month
      - name: day
        in: query
        required: true
        schema:
          type: integer
          description: Birth day
          title: Day
        description: Birth day
      - name: hour
        in: query
        required: true
        schema:
          type: integer
          description: Birth hour
          title: Hour
        description: Birth hour
      - name: minute
        in: query
        required: false
        schema:
          type: integer
          description: Birth minute (default 0)
          default: 0
          title: Minute
        description: Birth minute (default 0)
      - name: second
        in: query
        required: false
        schema:
          type: integer
          description: Birth second (optional, default 0)
          default: 0
          title: Second
        description: Birth second (optional, default 0)
      - name: place
        in: query
        required: true
        schema:
          type: string
          description: Birth place (city, country)
          title: Place
        description: Birth place (city, country)
      - name: transit_year
        in: query
        required: true
        schema:
          type: integer
          description: Transit year to analyze
          title: Transit Year
        description: Transit year to analyze
      - name: transit_month
        in: query
        required: true
        schema:
          type: integer
          description: Transit month to analyze
          title: Transit Month
        description: Transit month to analyze
      - name: transit_day
        in: query
        required: true
        schema:
          type: integer
          description: Transit day to analyze
          title: Transit Day
        description: Transit day to analyze
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /analyze/composite:
    post:
      summary: Analyze Composite
      description: "Calculate Composite Chart features for exactly 2 people.\nExposes detailed pairwise composite logic including new and duplicated channels/chakras.\n\nInput:\n{\n    \"person1\": { ... },\n    \"person2\": { ... }\n}"
      operationId: analyze_composite_analyze_composite_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties:
                $ref: '#/components/schemas/PersonInput'
              type: object
              title: Inputs
              description: Dictionary of exactly 2 people for Composite analysis.
            example:
              person1:
                place: Berlin, Germany
                year: 1985
                month: 6
                day: 15
                hour: 14
                minute: 30
              person2:
                place: Munich, Germany
                year: 1988
                month: 11
                day: 22
                hour: 9
                minute: 15
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /analyze/compmatrix:
    post:
      summary: Analyze Composite Matrix
      description: "Calculate Human Design features for multiple people and analyze\
        \ their composite combinations (matrix).\n\nInput:\n{\n    \"person1\": {\
        \ \"place\": \"City, Country\", \"year\": 1990, \"month\": 1, \"day\": 1,\
        \ \"hour\": 12, \"minute\": 0 },\n    \"person2\": { ... }\n}"
      operationId: analyze_compmatrix_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties:
                $ref: '#/components/schemas/PersonInput'
              type: object
              title: Inputs
              description: Dictionary of people where keys are names (e.g., person1,
                person2) and values are birth details.
            example:
              person1:
                place: Berlin, Germany
                year: 1985
                month: 6
                day: 15
                hour: 14
                minute: 30
              person2:
                place: Munich, Germany
                year: 1988
                month: 11
                day: 22
                hour: 9
                minute: 15
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /analyze/penta:
    post:
      summary: Analyze Penta
      description: "Calculate Penta (Group Analysis) for a group of people. Returns\
        \ the match percentage and details of active Penta gates.\n\nInput:\n{\n \
        \   \"person1\": { \"place\": \"Berlin, Germany\", \"year\": 1985, \"month\"\
        : 6, \"day\": 15, \"hour\": 14, \"minute\": 30 },\n    \"person2\": { ...\
        \ },\n    \"person3\": { ... }\n}"
      operationId: analyze_penta_analyze_penta_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties:
                $ref: '#/components/schemas/PersonInput'
              type: object
              title: Inputs
              description: Dictionary of 3-5 people for Penta analysis.
            example:
              person1:
                place: Berlin, Germany
                year: 1985
                month: 6
                day: 15
                hour: 14
                minute: 30
              person2:
                place: New York, USA
                year: 1980
                month: 2
                day: 10
                hour: 9
                minute: 15
              person3:
                place: London, UK
                year: 1990
                month: 12
                day: 5
                hour: 18
                minute: 45
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PersonInput:
      properties:
        place:
          type: string
          minLength: 1
          title: Place
          description: Place of birth (City, Country)
        year:
          anyOf:
          - type: integer
          - type: string
          title: Year
          description: Birth year (1800-2100)
        month:
          anyOf:
          - type: integer
          - type: string
          title: Month
          description: Birth month (1-12)
        day:
          anyOf:
          - type: integer
          - type: string
          title: Day
          description: Birth day (1-31)
        hour:
          anyOf:
          - type: integer
          - type: string
          title: Hour
          description: Birth hour (0-23)
        minute:
          anyOf:
          - type: integer
          - type: string
          title: Minute
          description: Birth minute (0-59)
      type: object
      required:
      - place
      - year
      - month
      - day
      - hour
      - minute
      title: PersonInput
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer